+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Swashbuckle now troll combo with Fury Swipes (both of them) and arcane orb
 - 3 Mana Regen custom abilities (Flesh Heap, Survival, Souls) have been disabled until their values can be readjusted
 - Disable an Essence Aura talent that doesnt work
+- Added cheat to cheat menu: True Sight (-gem)
 
 ## 10.8
 - Lifebreak Redux now troll combo with Dispersion